requirements.txt
pandas>=2.2.0,<3.0
numpy>=1.26.0
yfinance>=0.2.40
PyYAML>=6.0.1
requests>=2.31.0
streamlit>=1.37.0
